moved EntitlementPeriodType to Entitlment class
