+++ packages/nativescript-revenuecat/Entitlement/index.ios.ts
@@ -1,5 +1,20 @@
 import { BaseEntitlement } from './common';
 
+/**
+ * @type {RCPeriodType}
+ * @description * RevenueCat iOS SDK Entitlement period types.
+ *
+ * Native Enum value mapping:
+ * - Normal: 0
+ * - Intro: 1
+ * - Trial: 2
+ */
+export const EntitlementPeriodType = {
+  NORMAL: RCPeriodType.Normal,
+  INTRO: RCPeriodType.Intro,
+  TRIAL: RCPeriodType.Trial,
+};
+
 export class Entitlement extends BaseEntitlement {
   public nativeValue: RCEntitlementInfo;
