--- packages/nativescript-revenuecat/util/platform-utils.d.ts
@@ -1,1 +1,0 @@
-export declare const EntitlementPeriodType;
