--- packages/nativescript-revenuecat/util/platform-utils.ios.ts
@@ -1,14 +1,0 @@
-/**
- * @type {RCPeriodType}
- * @description * RevenueCat iOS SDK Entitlement period types.
- *
- * Native Enum value mapping:
- * - Normal: 0
- * - Intro: 1
- * - Trial: 2
- */
-export const EntitlementPeriodType = {
-  NORMAL: RCPeriodType.Normal,
-  INTRO: RCPeriodType.Intro,
-  TRIAL: RCPeriodType.Trial,
-};
